+++ src/utils/EventManager.ts
@@ -2,6 +2,7 @@ export type FeatureName =
 	| "automaticTheaterMode"
 	| "featureMenu"
 	| "hideScrollBar"
+	| "hideShorts"
 	| "loopButton"
 	| "maximizePlayerButton"
 	| "openTranscriptButton"
@@ -13,6 +14,7 @@ export type FeatureName =
 	| "screenshotButton"
 	| "scrollWheelSpeedControl"
 	| "scrollWheelVolumeControl"
+	| "shortsAutoScroll"
 	| "videoHistory"
 	| "volumeBoostButton";
 type EventCallback<K extends keyof HTMLElementEventMap> = (event: HTMLElementEventMap[K]) => void;